extended test coverage
