+++ system_tests/test_all.py
@@ -5,9 +5,11 @@
 import test_interlocking_examples
 import test_automation_examples
 import test_schematic_editor
+import test_object_editing
 import system_test_harness
 
 test_schematic_editor.run_all_schematic_editor_tests()
 test_interlocking_examples.run_all_interlocking_example_tests()
 test_automation_examples.run_all_automation_example_tests()
+test_object_editing.run_all_object_editing_tests()
 system_test_harness.complete_tests(shutdown=True)
