+++ system_tests/test_schematic_editor.py
@@ -12,15 +12,14 @@ from system_test_harness import *
 #    Delete objects
 #    Rotate objects
 #    Move (drag and drop) objects
-#    Move (drag and drop) Line ends
 #    Editing of Lines (Drag and drop line ends)
 #    copy and paste objects
 #-----------------------------------------------------------------------------------
 
 def run_basic_editor_tests(delay:float=0.0):
-    print("Bssic schematic editor tests")
+    print("Basic schematic editor tests")
     # Add elements to the layout and move them to their initial positions
-    s1 = create_colour_light_signal()
+    s1 = create_colour_light_signal()                 
     sleep(delay)
     select_and_move_objects(s1,100,100,delay=delay)
     sleep(delay)
@@ -276,7 +275,7 @@ def run_basic_editor_tests(delay:float=0.0):
     delete_selected_objects()
     sleep(delay)
     assert_objects_do_not_exist(s11,s12,s13,s14,p11,p12,t11,l11,l12)
-    assert_objects_exists(s1,s2,s3,s4,p1,p2,t1,l1,l2,i1)
+    assert_objects_exist(s1,s2,s3,s4,p1,p2,t1,l1,l2,i1)
     # Finally test copy/paste of block instrument
     select_single_object(i1)
     sleep(delay)
@@ -295,575 +294,196 @@ def run_basic_editor_tests(delay:float=0.0):
     # clean up
     sleep(delay)
     select_all_objects()
+    sleep(delay)
     delete_selected_objects()
     return()
 
 #-----------------------------------------------------------------------------------
-# These test the point chaining functions, specifically:
-#    Points in a chain should always remain synchronised with the parent
-#    Change of Item ID in an auto switched chain (chain remains synchronised)
-#    Deletion of an item (chain is updated to remove the deleted point)
+# These test the schematic undo and re-do functions, specifically:
+#    The ability to move backwards and forwards through the undo buffer
+#    Resetting of the undo buffer following a change
 #-----------------------------------------------------------------------------------
 
-def run_point_chaining_tests(delay:float=0.0):
-    print("Point 'auto-switch' chaining tests")
-    # Add elements to the layout
+def run_undo_and_redo_tests(delay:float=0.0):
+    print("Schematic Undo and Redo tests")
+    # Initialise the test harness to give us a 'new' schematic
     sleep(delay)
-    p1 = create_left_hand_point()
+    initialise_test_harness()
+    # Create an initial event
     sleep(delay)
-    p2 = create_left_hand_point()
+    p1 = create_left_hand_point()                         # event 1 #
+    assert_objects_exist(p1)
+    # Re-initialise the test harness to give us a 'new' schematic
     sleep(delay)
-    p3 = create_left_hand_point()
+    initialise_test_harness()
+    assert_objects_do_not_exist(p1)
+    # Create a new chain of events
     sleep(delay)
-    p4 = create_left_hand_point()
-    assert_object_configuration(p1,{"itemid":1})
-    assert_object_configuration(p2,{"itemid":2})
-    assert_object_configuration(p3,{"itemid":3})
-    assert_object_configuration(p4,{"itemid":4})
-    # Set up a chain of points switching other points (P1 => P2 => P3 => P4)
-    update_object_configuration(p4,{"automatic":True})
-    update_object_configuration(p3,{"alsoswitch":4,"automatic":True})
-    update_object_configuration(p2,{"alsoswitch":3,"automatic":True})
-    update_object_configuration(p1,{"alsoswitch":2})
-    # Test the auto switch is working
+    s1 = create_colour_light_signal()                     # event 1 #
+    assert_objects_exist(s1)
     sleep(delay)
-    assert_points_normal(1,2,3,4)                            
-    set_points_switched(1)
-    assert_points_switched(1,2,3,4)
-    # Update the ID for P3
+    assert_object_position(s1,50,50)
+    select_and_move_objects(s1,100,100,delay=delay)       # event 2 #
+    assert_object_position(s1,100,100)
     sleep(delay)
-    assert_object_configuration(p2,{"alsoswitch":3})
-    update_object_configuration(p3,{"itemid":13})
-    assert_object_configuration(p3,{"itemid":13})
-    assert_object_configuration(p2,{"alsoswitch":13})
-    # As P3 is switched by P2 and P2 is switched, P3 should be re-created in
-    # the switched state and therefore all points should remain switched
-    assert_points_switched(1,2,13,4)
-    # Test the switching chain still works with the changed ID
+    l1 = create_line()                                    # event 3 #
+    assert_objects_exist(l1)
     sleep(delay)
-    set_points_normal(1)
-    assert_points_normal(1,2,13,4)
+    assert_object_position(l1,50,50,100,50)
+    select_and_move_objects(l1,500,150,delay=delay)       # event 4 #
+    assert_object_position(l1,475,150,525,150)
     sleep(delay)
-    set_points_switched(1)
-    assert_points_switched(1,2,13,4)
-    # Update the ID for P1
+    select_and_move_line_end(l1,1,25,100,delay=delay)     # event 5 #
+    assert_object_position(l1,25,100,525,150)
     sleep(delay)
-    update_object_configuration(p1,{"itemid":11})
-    assert_object_configuration(p1,{"itemid":11})
-    # As P1 is the start of the chain it should be created in its default
-    # 'normal' state and P2/P3 should be reverted to their normal state
-    assert_points_normal(11,2,13,4)
-    # Test that references are correctly removed on point deletion
+    select_and_move_line_end(l1,2,550,100,delay=delay)    # event 6 #
+    assert_object_position(l1,25,100,550,100)
     sleep(delay)
-    assert_object_configuration(p1,{"alsoswitch":2})
-    select_single_object(p2)
-    delete_selected_objects()
-    assert_object_configuration(p1,{"alsoswitch":0})
-    # Now the switching chain has been broken
+    select_single_object(s1)    
     sleep(delay)
-    set_points_switched(11)
-    assert_points_switched(11)
-    assert_points_normal(13,4)
-    # Make P3 a manual point and check the P3=>P4 chaining still works
+    assert_object_configuration(s1,{"orientation":0})  
+    rotate_selected_objects()                             # event 7 #
+    assert_object_configuration(s1,{"orientation":180})  
     sleep(delay)
-    update_object_configuration(p3,{"automatic":False})
-    set_points_switched(13)
-    assert_points_switched(13,4)
+    copy_selected_objects()
+    [s11] = paste_clipboard_objects()                     # event 8 #
+    assert_objects_exist(s11)
     sleep(delay)
-    set_points_normal(13)
-    assert_points_normal(13,4)
-    # clean up
+    delete_selected_objects()                             # event 9 #
+    assert_objects_do_not_exist(s11)
     sleep(delay)
-    select_all_objects()
-    delete_selected_objects()
-    return()
-
-#-----------------------------------------------------------------------------------
-# These test the mirrored sections functions, specifically:
-#    Mirrored sections in a chain should be updated with any changes to the parent
-#    Change of Item ID in an mirrored section chain (chain remains synchronised)
-#    Deletion of an item (chain is updated to remove the deleted section)
-#    Use case of two sections mirroring each other - should remain synchronised
-#-----------------------------------------------------------------------------------
-
-def run_mirrored_section_tests(delay:float=0.0):
-    print("Mirrored Section tests")
-    t1 = create_track_section()
+    assert_object_configuration(s1,{"itemid":1})  
+    update_object_configuration(s1,{"itemid":2})          # event 10 #
+    assert_object_configuration(s1,{"itemid":2})
+    # check undo
     sleep(delay)
-    t2 = create_track_section()
+    assert_object_configuration(s1,{"itemid":2})
+    undo()
+    assert_object_configuration(s1,{"itemid":1})          # undo of event 10 #
     sleep(delay)
-    t3 = create_track_section()
+    assert_objects_do_not_exist(s11)                 
+    undo()                                                # undo of event 9 #
+    assert_objects_exist(s11)
     sleep(delay)
-    t4 = create_track_section()
-    # Check the default item IDs have been assigned correctly
-    assert_object_configuration(t1,{"itemid":1})
-    assert_object_configuration(t2,{"itemid":2})
-    assert_object_configuration(t3,{"itemid":3})
-    assert_object_configuration(t4,{"itemid":4})
-    # Set up a chain of Mirrored sections (P1 => P2 => P3 => P4)
-    # Note that section IDs are strings rather than integers
-    update_object_configuration(t4,{"mirror":'3'})
-    update_object_configuration(t3,{"mirror":'2'})
-    update_object_configuration(t2,{"mirror":'1'})
-    # Test the basic mirroring is working - Note that updating
-    # mirrored is not recursive - When a change is made to one
-    # section then only the 'mirrored' section is updated.
-    # Changes are not propogated further if that section is
-    # also 'mirrored' by another section
+    undo()                                                # undo of event 8 #
+    assert_objects_do_not_exist(s11)
     sleep(delay)
-    set_run_mode()
-    assert_sections_clear(1,2,3,4)
+    assert_object_configuration(s1,{"orientation":180})
+    undo()                                                # undo of event 7 #
+    assert_object_configuration(s1,{"orientation":0})
     sleep(delay)
-    # Section 1 is mirrored by section 2
-    set_sections_occupied(1)
-    assert_sections_occupied(1,2,3,4)
+    assert_object_position(l1,25,100,550,100)
+    undo()                                                # undo of event 6 #
+    assert_object_position(l1,25,100,525,150)
     sleep(delay)
-    set_sections_clear(1)
-    assert_sections_clear(1,2,3,4)
-    # Section 2 is mirrored by section 3
+    assert_object_position(l1,25,100,525,150)
+    undo()                                                # undo of event 5 #
+    assert_object_position(l1,475,150,525,150)
     sleep(delay)
-    set_sections_occupied(2)
-    assert_sections_occupied(2,3,4)
-    assert_sections_clear(1)
+    undo()                                                # undo of event 4 #
+    assert_object_position(l1,50,50,100,50)    
     sleep(delay)
-    set_sections_clear(2)
-    assert_sections_clear(1,2,3,4)
-    # Section 3 is mirrored by section 2
+    assert_objects_exist(l1)
+    undo()                                                # undo of event 3 #
+    assert_objects_do_not_exist(l1)    
     sleep(delay)
-    set_sections_occupied(3)
-    assert_sections_occupied(3,4)
-    assert_sections_clear(1,2)
+    assert_object_position(s1,100,100)
+    undo()                                                # undo of event 2 #
+    assert_object_position(s1,50,50)
     sleep(delay)
-    set_sections_clear(3)
-    assert_sections_clear(1,2,3,4)
-    # Update the ID for T3
+    assert_objects_exist(s1)
+    undo()                                                # undo of event 1 #
+    assert_objects_do_not_exist(s1)
+    # Test we can't go back any further (buffer was reset)
+    undo()                                               
+    assert_objects_do_not_exist(p1)
+    undo()                                               
+    assert_objects_do_not_exist(p1)
+    undo()                                               
+    assert_objects_do_not_exist(p1)
+    undo()                                               
+    assert_objects_do_not_exist(p1)
+    # Test re-do of all 10 events
     sleep(delay)
-    set_edit_mode()
-    update_object_configuration(t3,{"itemid":13})
-    assert_object_configuration(t3,{"itemid":13})
-    assert_object_configuration(t4,{"mirror":'13'})
+    assert_objects_do_not_exist(s1)
+    redo()
+    assert_objects_exist(s1)
     sleep(delay)
-    set_run_mode()
-    # Test the mirroring still works
+    assert_object_position(s1,50,50)
+    redo()
+    assert_object_position(s1,100,100)
     sleep(delay)
-    set_sections_occupied(1)
-    assert_sections_occupied(1,2,13,4)
+    assert_objects_do_not_exist(l1)
+    redo()
+    assert_objects_exist(l1)
     sleep(delay)
-    set_sections_clear(1)
-    assert_sections_clear(1,2,13,4)
-    # Delete T3
+    assert_object_position(l1,50,50,100,50)
+    redo()
+    assert_object_position(l1,475,150,525,150)
     sleep(delay)
-    set_edit_mode()
-    select_single_object(t3)
-    delete_selected_objects()
-    assert_object_configuration(t4,{"mirror":''})
+    redo()
+    assert_object_position(l1,25,100,525,150)
     sleep(delay)
-    set_run_mode()
-    # The mirrored chain should be broken
+    redo()
+    assert_object_position(l1,25,100,550,100)
     sleep(delay)
-    set_sections_occupied(1)
-    assert_sections_occupied(1,2)
-    assert_sections_clear(4)
+    select_single_object(s1)    
     sleep(delay)
-    set_sections_clear(1)
-    assert_sections_clear(1,2,4)
-    # Setup sections 1 and 2 to mirror each other
+    assert_object_configuration(s1,{"orientation":0})  
+    redo()
+    assert_object_configuration(s1,{"orientation":180})  
     sleep(delay)
-    set_edit_mode()
-    update_object_configuration(t1,{"mirror":'2'})
+    copy_selected_objects()
+    assert_objects_do_not_exist(s11)
+    redo()
+    assert_objects_exist(s11)
     sleep(delay)
-    set_run_mode()
-    # Section 1 is mirrored by section 2
+    redo()
+    assert_objects_do_not_exist(s11)
     sleep(delay)
-    set_sections_occupied(1)
-    assert_sections_occupied(1,2)
+    assert_object_configuration(s1,{"itemid":1})  
+    redo()
+    assert_object_configuration(s1,{"itemid":2})   
+    # Test we can't go any further forward (end of buffer)
+    redo()                                               
+    redo()                                               
+    redo()                                               
+    redo()
+    # Test undo again
     sleep(delay)
-    set_sections_clear(1)
-    assert_sections_clear(1,2)
-    # Section 2 is mirrored by section 1
-    sleep(delay)
-    set_sections_occupied(2)
-    assert_sections_occupied(1,2)
-    sleep(delay)
-    set_sections_clear(2)
-    assert_sections_clear(1,2)
-    # clean up
-    sleep(delay)
-    set_edit_mode()
-    sleep(delay)
-    select_all_objects()
-    delete_selected_objects()
-    return()
-
-#-----------------------------------------------------------------------------------
-# These test the changes between Run and edit mode, specifically:
-#    Section state from run mode is 'remembered' during edit mode
-#-----------------------------------------------------------------------------------
-
-def run_mode_change_tests(delay:float=0.0):
-    print("Mode Change tests")
-    # Add elements to the layout
-    sleep(delay)
-    s1 = create_colour_light_signal()
-    sleep(delay)
-    t1 = create_track_section()
-    sleep(delay)
+    assert_object_configuration(s1,{"itemid":2})
+    undo()                                                # undo of event 10 #
     assert_object_configuration(s1,{"itemid":1})
-    assert_object_configuration(t1,{"itemid":1})
-    # Configure Signal 1 to be overridden if Section 1 is occupied
-    update_object_configuration(s1,{
-                "tracksections":[0,[1,0,0,0,0]],
-                "overridesignal":True})
-    # Test the override works
     sleep(delay)
-    set_run_mode()
-    set_signals_off(1)
-    assert_signals_override_clear(1)
-    assert_signals_PROCEED(1)
+    assert_objects_do_not_exist(s11)                 
+    undo()                                                # undo of event 9 #
+    assert_objects_exist(s11)
+    # make a change to delete the buffer of re-dos from the current point
     sleep(delay)
-    set_sections_occupied(1)
-    assert_signals_override_set(1)
-    assert_signals_DANGER(1)
-    # Set edit mode - Overrides are not maintained in Edit mode 
-    sleep(delay)
-    set_edit_mode()
-    assert_signals_override_clear(1)
-    assert_signals_PROCEED(1)
-    # Return to Run mode - Overrides should be re-instated
-    sleep(delay)
-    set_run_mode()
-    assert_signals_override_set(1)
-    assert_signals_DANGER(1)
-    # Clear the section (and the Override)
-    sleep(delay)
-    set_sections_clear(1)
-    assert_signals_override_clear(1)
-    assert_signals_PROCEED(1)
-    # Set edit trhen run mode - section remains clear
-    sleep(delay)
-    set_edit_mode()
-    sleep(delay)
-    set_run_mode()
-    assert_signals_override_clear(1)
-    assert_signals_PROCEED(1)
-    # clean up
-    sleep(delay)
-    set_edit_mode()
-    sleep(delay)
-    select_all_objects()
-    delete_selected_objects()
-    return()
-    
-#-----------------------------------------------------------------------------------
-# These test the Item ID update functions, specifically:
-#    Deletion of an item (chain is updated to remove the deleted point
-#    Update of point interlocking tables when Signal ID is changed or deleted
-#    Update of instrument interlocking tables when Signal ID is changed or deleted
-#    Update of signal interlocking tables when Signal ID is changed or deleted
-#    Update of signal interlocking tables when Point ID is changed or deleted
-#    Update of signal interlocking tables when Instrument ID is changed or deleted
-#    Update of signal automation tables when Section ID is changed or deleted
-#-----------------------------------------------------------------------------------
-
-def run_change_of_item_id_tests(delay:float=0.0):
-    print("Change of Item Id tests")
-    # Add elements to the layout
-    sleep(delay)
-    p1 = create_left_hand_point()
-    sleep(delay)
-    p2 = create_left_hand_point()
-    sleep(delay)
-    s1 = create_semaphore_signal()
-    sleep(delay)
-    s2 = create_semaphore_signal()
-    sleep(delay)
-    s3 = create_semaphore_signal()
-    sleep(delay)
-    t1 = create_track_section()
-    sleep(delay)
-    i1 = create_block_instrument()
-    # Test the 'one-up' IDs have been correctly generated
-    assert_object_configuration(p1,{"itemid":1})
-    assert_object_configuration(p2,{"itemid":2})
+    p1 = create_left_hand_point()                        
+    # Test we can't go any further forward (end of buffer)
+    redo()                                               
+    assert_objects_exist(s11)                 
     assert_object_configuration(s1,{"itemid":1})
-    assert_object_configuration(s2,{"itemid":2})
-    assert_object_configuration(s3,{"itemid":3})
-    assert_object_configuration(i1,{"itemid":1})
-    assert_object_configuration(t1,{"itemid":1})
+    redo()                                               
+    assert_objects_exist(s11)                 
+    assert_object_configuration(s1,{"itemid":1})
+    redo()                                               
+    assert_objects_exist(s11)                 
+    assert_object_configuration(s1,{"itemid":1})
+    redo()                                               
+    assert_objects_exist(s11)                 
+    assert_object_configuration(s1,{"itemid":1})
+    # one final undo
     sleep(delay)
-    select_and_move_objects(i1,500,200,delay=delay)
-    # Set up the signal interlocking tables and the automation tables for Signal 1
-    # This includes point/signal/inst interlocking, track sections & timed signals
-    # Signal is interlocked on instrument 1 and points 1/2 - Sig ahead is 2
-    update_object_configuration(s1,{
-        "sigroutes":[True,True,True,True,True],
-        "pointinterlock":[
-           [[[1,True],[1,True],[1,True],[1,True],[2,True],[2,True],[2,True]],"2",1],
-           [[[1,True],[1,True],[1,True],[1,True],[2,True],[2,True],[2,True]],"2",1],
-           [[[1,True],[1,True],[1,True],[1,True],[2,True],[2,True],[2,True]],"2",1],
-           [[[1,True],[1,True],[1,True],[1,True],[2,True],[2,True],[2,True]],"2",1],
-           [[[1,True],[1,True],[1,True],[1,True],[2,True],[2,True],[2,True]],"2",1] ],
-        "siginterlock":[
-             [ [2, [True, True, True, True, True]], 
-               [2, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]] ], 
-             [ [2, [True, True, True, True, True]], 
-               [2, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]] ], 
-             [ [2, [True, True, True, True, True]], 
-               [2, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]] ], 
-             [ [2, [True, True, True, True, True]], 
-               [2, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]] ], 
-             [ [2, [True, True, True, True, True]], 
-               [2, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]], 
-               [3, [True, True, True, True, True]] ] ],
-        "tracksections":[1, [1,1,1,1,1] ],
-        "timedsequences":[ [True,2,1,1],
-                           [True,2,1,1],
-                           [True,3,1,1],
-                           [True,2,1,1],
-                           [True,2,1,1] ]} )
-    # Assert the point interlocking tables have been configured correctly
-    assert_object_configuration(p1,{
-        "siginterlock":[ [1, [True,True,True,True,True] ] ] } )
-    assert_object_configuration(p2,{
-        "siginterlock":[ [1, [True,True,True,True,True] ] ] } )
-    # Configure another signal for point interlocking
-    update_object_configuration(s2,{
-        "sigroutes":[True,True,True,True,True],
-        "pointinterlock":[
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0],
-           [[[1,True],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0],
-           [[[1,True],[0,False],[0,False],[0,True],[0,False],[0,False],[0,False]],"",0],
-           [[[1,True],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0],
-           [[[1,True],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0] ] } )
-    # Assert the point interlocking tables have been configured correctly
-    assert_object_configuration(p1,{
-        "siginterlock":[ [1, [True,True,True,True,True] ],
-                         [2, [False,True,True,True,True] ] ] } )
-    assert_object_configuration(p2,{
-        "siginterlock":[ [1, [True,True,True,True,True] ] ] } )
-
-    # Test the basic interlocking
-    ################ ToDo - Interlocking with Block Instrument ########################
-    assert_signals_locked(1)
-    assert_points_unlocked(1,2)
-    sleep(delay)
-    set_points_switched(1,2)
-    assert_signals_unlocked(1)
-    sleep(delay)
-    set_signals_off(1)
-    assert_points_locked(1,2)
-    sleep(delay)
-    set_signals_on(1)
-    sleep(delay)
-    set_points_normal(1,2)
-    assert_signals_locked(1)    
-    # Change the IDs of Signal 2, Points 1/2, Instrument 1 and Track Section 1 
-    update_object_configuration(s1,{"itemid":11})
-    update_object_configuration(s2,{"itemid":12})
-    update_object_configuration(s3,{"itemid":13})
-    update_object_configuration(p1,{"itemid":21})
-    update_object_configuration(p2,{"itemid":22})
-    update_object_configuration(t1,{"itemid":31})
-    update_object_configuration(i1,{"itemid":41})
-    # Test the IDs have been changed
-    assert_object_configuration(s1,{"itemid":11})
-    assert_object_configuration(s2,{"itemid":12})
-    assert_object_configuration(s3,{"itemid":13})
-    assert_object_configuration(p1,{"itemid":21})
-    assert_object_configuration(p2,{"itemid":22})
-    assert_object_configuration(t1,{"itemid":31})
-    assert_object_configuration(i1,{"itemid":41})
-    # Test the interlocking and automation tables have been updated
-    assert_object_configuration(s1,{
-        "pointinterlock":[
-           [[[21,True],[21,True],[21,True],[21,True],[22,True],[22,True],[22,True]],"12",41],
-           [[[21,True],[21,True],[21,True],[21,True],[22,True],[22,True],[22,True]],"12",41],
-           [[[21,True],[21,True],[21,True],[21,True],[22,True],[22,True],[22,True]],"12",41],
-           [[[21,True],[21,True],[21,True],[21,True],[22,True],[22,True],[22,True]],"12",41],
-           [[[21,True],[21,True],[21,True],[21,True],[22,True],[22,True],[22,True]],"12",41] ],
-        "siginterlock":[
-             [ [12, [True, True, True, True, True]], 
-               [12, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]] ], 
-             [ [12, [True, True, True, True, True]], 
-               [12, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]] ], 
-             [ [12, [True, True, True, True, True]], 
-               [12, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]] ], 
-             [ [12, [True, True, True, True, True]], 
-               [12, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]] ], 
-             [ [12, [True, True, True, True, True]], 
-               [12, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]] ] ],
-        "tracksections":[31,[31,31,31,31,31] ],
-        "timedsequences":[ [True,12,1,1],
-                           [True,12,1,1],
-                           [True,13,1,1],
-                           [True,12,1,1],
-                           [True,12,1,1] ]} )
-    # Assert the point interlocking tables have been configured correctly
-    assert_object_configuration(p1,{
-        "siginterlock":[ [11, [True,True,True,True,True] ],
-                         [12, [False,True,True,True,True] ] ] } )
-    assert_object_configuration(p2,{
-        "siginterlock":[ [11, [True,True,True,True,True] ] ] } )
-    # Test the basic interlocking
-    ############ ToDo - Interlocking with Block Instrument ########################
-    assert_signals_locked(11)
-    assert_points_unlocked(21,22)
-    sleep(delay)
-    set_points_switched(21,22)
-    assert_signals_unlocked(11)
-    sleep(delay)
-    set_signals_off(11)
-    assert_points_locked(21,22)
-    sleep(delay)
-    set_signals_on(11)
-    sleep(delay)
-    set_points_normal(21,22)
-    assert_signals_locked(11)
-    # Delete Point 21 and test all references have been removed
-    # Note that the point 22 entries will have shuffled down in the list
-    sleep(delay)
-    select_single_object(p1)
-    sleep(delay)
-    delete_selected_objects()
-    assert_object_configuration(s1,{
-        "pointinterlock":[
-           [[[22,True],[22,True],[22,True],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[22,True],[22,True],[22,True],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[22,True],[22,True],[22,True],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[22,True],[22,True],[22,True],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[22,True],[22,True],[22,True],[0,False],[0,False],[0,False],[0,False]],"12",41] ] })
-    # Delete Point 22 and test all references have been removed
-    sleep(delay)
-    select_single_object(p2)
-    sleep(delay)
-    delete_selected_objects()
-    assert_object_configuration(s1,{
-        "pointinterlock":[
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"12",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"12",41] ] })
-    # Delete signal 12 and test all references have been removed
-    sleep(delay)
-    select_single_object(s2)
-    sleep(delay)
-    delete_selected_objects()
-    assert_object_configuration(s1,{
-        "pointinterlock":[
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",41],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",41] ],
-        "siginterlock":[
-             [ [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [13, [True, True, True, True, True]], 
-               [13, [True, True, True, True, True]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ] ],
-        "timedsequences":[ [False,0,0,0],
-                           [False,0,0,0],
-                           [True,13,1,1],
-                           [False,0,0,0],
-                           [False,0,0,0] ] } )
-    # Delete Instrument 31 and test all references have been removed
-    sleep(delay)
-    select_single_object(i1)
-    sleep(delay)
-    delete_selected_objects()
-    assert_object_configuration(s1,{
-        "pointinterlock":[
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0],
-           [[[0,False],[0,False],[0,False],[0,False],[0,False],[0,False],[0,False]],"",0] ] })
-    # Delete Section 41 and test all references have been removed
-    sleep(delay)
-    select_single_object(t1)
-    sleep(delay)
-    delete_selected_objects()
-    assert_object_configuration(s1,{
-        "tracksections":[0,[0,0,0,0,0] ] } )
-    # Delete signal 13 and test all references have been removed
-    sleep(delay)
-    select_single_object(s3)
-    sleep(delay)
-    delete_selected_objects()
-    assert_object_configuration(s1,{
-        "siginterlock":[
-             [ [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ], 
-             [ [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]], 
-               [0, [False, False, False, False, False]] ] ],
-        "timedsequences":[ [False,0,0,0],
-                           [False,0,0,0],
-                           [False,0,0,0],
-                           [False,0,0,0],
-                           [False,0,0,0] ] } )
+    undo()
+    assert_objects_do_not_exist(p1)
     # clean up
     sleep(delay)
     select_all_objects()
+    sleep(delay)
     delete_selected_objects()
     return()
-
-#-----------------------------------------------------------------------------------
-# ToDo:
-#    Undo and Redo
-#    Reset All Objects
-#    
-#-----------------------------------------------------------------------------------
 
 ######################################################################################################
 
@@ -871,10 +491,7 @@ def run_all_schematic_editor_tests(delay=0):
     initialise_test_harness()
     set_edit_mode()
     run_basic_editor_tests(delay)
-    run_point_chaining_tests(delay)
-    run_mirrored_section_tests(delay)
-    run_mode_change_tests(delay)
-    run_change_of_item_id_tests(delay)
+    run_undo_and_redo_tests(delay)
     
 if __name__ == "__main__":
     run_all_schematic_editor_tests(delay = 0.5)
